fixed some constants to be u8

+++ sources/runtime.c
@@ -5,12 +5,12 @@ static __sfr __at 0xfe PortFE;
 
 void border(u8 colour)
 {
-	_last_value = (_last_value & 0xf8) | colour;
+	_last_value = (_last_value & (u8)0xf8) | colour;
 	PortFE = _last_value;
 }
 
 void speaker(u8 bit)
 {
-	_last_value = (_last_value & 0xef) | (bit? 0x10: 0x00);
+	_last_value = (_last_value & (u8)0xef) | (bit? (u8)0x10: (u8)0x00);
 	PortFE = _last_value;
 }
